+++ entities/db/connection.go
@@ -13,7 +13,7 @@ var pool *redis.Pool
 
 // This function is called in order to insure propper db acsess.
 // It creates the DB connection and stores it in the connection variable.
-func InitPool(host string, port int) {
+func InitPool(host string, port uint16) {
 	log.Print("Initializing database connection... ")
 	serverAddr := fmt.Sprintf("%v:%v", host, port)
